+++ R/typical.R
@@ -5,21 +5,35 @@
 #'
 #'
 #' @param x A variable.
-#' @param fun Character vector, naming the function to be applied to numeric
+#' @param fun Character vector, naming the function to be applied to
 #'        \code{x}. Currently, \code{"mean"}, \code{"weighted.mean"},
 #'        \code{"median"} and \code{"mode"} are supported, which call the
 #'        corresponding R functions (except \code{"mode"}, which calls an
 #'        internal function to compute the most common value). \code{"zero"}
-#'        simply returns 0.
+#'        simply returns 0. \strong{Note:} By default, if \code{x} is a factor,
+#'        only \code{fun = "mode"} is applicable; for all other values (including
+#'        the default, \code{"mean"}) the reference level of \code{x} is returned.
+#'        For character vectors, only the mode is returned. You can use a named
+#'        vector to apply other different functions to numeric and categorical
+#'        \code{x}, where factor are first converted to numeric vectors, e.g.
+#'        \code{fun = c(numeric = "median", factor = "mean")}. See 'Examples'.
 #' @param ... Further arguments, passed down to \code{fun}.
 #'
 #' @return The "typical" value of \code{x}.
 #'
 #' @details By default, for numeric variables, \code{typical_value()} returns the
 #'          mean value of \code{x} (unless changed with the \code{fun}-argument).
-#'          For factors, the reference level and for character vectors, to the
-#'          most common value (mode) is returned.
-
+#'          \cr \cr
+#'          For factors, the reference level is returned or the most common value
+#'          (if \code{fun = "mode"}), unless \code{fun} is a named vector. If
+#'          \code{fun} is a named vector, specify the function for numeric
+#'          and categorical variables as element names, e.g.
+#'          \code{fun = c(numeric = "median", factor = "mean")}. In this case,
+#'          factors are converted to numeric values (using \code{\link[sjmisc]{to_value}})
+#'          and the related function is applied. You may abbreviate the names
+#'          \code{fun = c(n = "median", f = "mean")}. See also 'Examples'.
+#'          \cr \cr
+#'          For character vectors the most common value (mode) is returned.
 #'
 #' @examples
 #' data(iris)
@@ -35,9 +49,37 @@
 #' typical_value(x, "weighted.mean")
 #' typical_value(x, "weighted.mean", w = wt)
 #'
+#' # for factors, return either reference level or mode value
+#' set.seed(123)
+#' x <- sample(iris$Species, size = 30, replace = TRUE)
+#' typical_value(x)
+#' typical_value(x, fun = "mode")
+#'
+#' # for factors, use a named vector to apply other functions than "mode"
+#' map(iris, ~ typical_value(.x, fun = c(n = "median", f = "mean")))
+#'
+#'
 #' @export
-typical_value <- function(x, fun = c("mean", "median", "mode", "weighted.mean", "zero"), ...) {
-  fun <- match.arg(fun)
+typical_value <- function(x, fun = "mean", ...) {
+
+  # check if we have named vectors and find the requested function
+  # for special functions for factors, convert to numeric first
+
+  fnames <- names(fun)
+
+  if (!is.null(fnames)) {
+    if (is.numeric(x)) {
+      fun <- fun[which(fnames %in% c("numeric", "n"))]
+    } else if (is.factor(x)) {
+      fun <- fun[which(fnames %in% c("factor", "f"))]
+      x <- sjmisc::to_value(x, keep.labels = FALSE)
+    }
+  }
+
+
+  if (!fun %in% c("mean", "median", "mode", "weighted.mean", "zero"))
+    stop("`fun` must be one of \"mean\", \"median\", \"mode\", \"weighted.mean\" or \"zero\".", call. = F)
+
 
   if (fun == "median")
     myfun <- get("median", asNamespace("stats"))
@@ -50,12 +92,16 @@ typical_value <- function(x, fun = c("mean", "median", "mode", "weighted.mean", 
   else
     myfun <- get("mean", asNamespace("base"))
 
-  if (is.numeric(x))
+  if (is.numeric(x)) {
     do.call(myfun, args = list(x = x, na.rm = TRUE, ...))
-  else if (is.factor(x))
-    levels(x)[1]
-  else
+  } else if (is.factor(x)) {
+    if (fun != "mode")
+      levels(x)[1]
+    else
+      mode_value(x)
+  } else {
     mode_value(x)
+  }
 }
 
 
@@ -64,8 +110,11 @@ mode_value <- function(x, ...) {
   counts <- table(x)
   modus <- names(counts)[max(counts) == counts]
 
+  # in case all values appear equally often, use first value
+  if (length(modus) > 1) modus <- modus[1]
+
   # check if it's numeric
-  if (!is.na(as.numeric(modus)))
+  if (!is.na(suppressWarnings(as.numeric(modus))))
     as.numeric(modus)
   else
     modus